requirements.txt
backpack_for_pytorch==1.6.0
numpy==1.24.4
openfermion==1.6.1
openfermionpsi4==0.5
PubChemPy==1.0.4
pytorch_model_summary==0.1.2
scipy==1.13.0
tensorboardX==2.6.2.2
torch==1.12.0
tqdm==4.66.1
yacs==0.1.8
